fix broken ct_int integration test
