+++ test/integration/cluster_solver/stdthread_qmci/stdthread_ctint_test_per_sample_output.json
@@ -49,7 +49,7 @@
         {
                 "warm-up-sweeps"         : 100,
                 "sweeps-per-measurement" : 1,
-                "measurements"           : 400,
+                "measurements"           : 20,
 
                 "threaded-solver" : {
                         "walkers": 1,
